+++ src/bask_export.h
@@ -3,10 +3,10 @@
 
 #define EXPORT_CSV_SPLIT ";"
 
-int export_baskbin (bask_core*, struct bask_task**, char*);
-int export_csv (bask_core*, struct bask_task**, char*);
-int export_ical_getdatestr (char*, char*);
-int export_ical (bask_core*, struct bask_task**, char*);
-int export_web (bask_core*, struct bask_task**, char*);
+short export_baskbin (bask_core*, struct bask_task**, char*);
+short export_csv (bask_core*, struct bask_task**, char*);
+short export_ical_getdatestr (char*, char*);
+short export_ical (bask_core*, struct bask_task**, char*);
+short export_web (bask_core*, struct bask_task**, char*);
 
 #endif /* __BASK_EXPORT_H__ */
